[feature/6-create-refactoring-project-member] UserCrewDetailResponseDto 경로 변경
